04-llm-agents/financial_analyst/agent.py
import os
from google.adk.agents.llm_agent import Agent
from dotenv import load_dotenv

load_dotenv()

root_agent = Agent(
    model=os.getenv("GEMINI_MODEL", "gemini-2.5-flash"),
    name="fin_analyst",
    description="A professional financial analyst who interprets market data.",
    instruction=(
        "You are a Junior Financial Analyst at a major investment bank. "
        "Your job is to interpret financial data, 10-K filings, and earnings reports for clients. "
        "When asked about a stock (e.g., share price, PE ratio), provide a concise summary. "
        "If asked about earnings or 10-K risks, highlight the top 3 key takeaways. "
        "Maintain a professional, objective tone and always include a standard disclaimer that this is not investment advice."
    )
)
